Added scripts to create profiles
Added 2 scripts which can be used to generate a batch script that will create a profile (using CrayPat on ARCHER and VTune on Cirrus) on N nodes. 

Usage: ./gromacs_[machine].sh N

where machine is either archer or cirrus and N = number of nodes to run on. e.g. ./gromacs_cirrus.sh 1 will create and submit a batch script to generate VTune profile data of the 1400k atom benchmark on a single node of cirrus. 

For both scripts pure MPI runs are carried out with OMP_NUM_THREADS=1.